Extract storage layer into flashq-storage crate
• Create flashq-storage crate with storage backends and Record types
• Move memory and file storage implementations to flashq-storage
• Relocate 17 integration tests and 3 benchmarks to flashq-storage
• Update flashq to re-export storage types for backward compatibility
• Update flashq-grpc and flashq-cluster dependencies
• Update documentation to reflect new crate structure
