+++ Cargo.toml
@@ -2,7 +2,7 @@
 members = [
     "crates/flashq",
     "crates/flashq-grpc",
-    "crates/flashq-cluster",
+    "crates/flashq-cluster", "crates/flashq-storage",
 ]
 # Build only the core crate by default so `cargo build` remains lean
 default-members = ["crates/flashq"]
