+++ crates/flashq-cluster/src/lib.rs
@@ -36,7 +36,7 @@ pub use service::ClusterServiceImpl;
 // Re-export core flashq types and services for flashq-grpc
 pub use flashq::{FlashQ, Record, RecordWithOffset};
 pub mod storage {
-    pub use flashq::storage::*;
+    pub use flashq::*;
 }
 
 // Re-export logging macros for consistent usage across the crate
