+++ crates/flashq-storage/Cargo.toml
@@ -1,0 +1,42 @@
+[package]
+name = "flashq-storage"
+edition.workspace = true
+version.workspace = true
+license.workspace = true
+
+[lib]
+path = "src/lib.rs"
+
+[dependencies]
+serde.workspace = true
+serde_json.workspace = true
+chrono.workspace = true
+dashmap.workspace = true
+parking_lot.workspace = true
+fs4.workspace = true
+sysinfo.workspace = true
+log.workspace = true
+tracing.workspace = true
+libc = "0.2"
+
+[dev-dependencies]
+flashq = { path = "../flashq" }
+uuid.workspace = true
+tempfile.workspace = true
+divan = "0.1"
+test-log.workspace = true
+
+[[bench]]
+name = "memory_storage"
+harness = false
+path = "benches/memory_storage.rs"
+
+[[bench]]
+name = "file_storage"
+harness = false
+path = "benches/file_storage.rs"
+
+[[bench]]
+name = "batching_baseline"
+harness = false
+path = "benches/batching_baseline.rs"
