+++ crates/flashq-storage/benches/batching_baseline.rs
@@ -1,6 +1,6 @@
 use divan::{AllocProfiler, Bencher, black_box};
-use flashq::storage::{StorageBackend, file::SyncMode};
 use flashq::{FlashQ, Record};
+use flashq_storage::{StorageBackend, file::SyncMode};
 use std::collections::HashMap;
 
 #[global_allocator]
