+++ crates/flashq-storage/src/storage/backend.rs
@@ -4,9 +4,8 @@ use crate::storage::{
     ConsumerGroup, ConsumerOffsetStore, InMemoryConsumerGroup, InMemoryConsumerOffsetStore,
     InMemoryTopicLog, TopicLog,
 };
-use crate::warn;
 use fs4::fs_std::FileExt;
-use log::debug;
+use log::{debug, warn};
 use parking_lot::RwLock;
 use std::fs::{File, OpenOptions};
 use std::io::Write;
