+++ crates/flashq-storage/src/storage/trait.rs
