+++ crates/flashq-storage/tests/storage/batching_tests.rs
@@ -1,6 +1,6 @@
-use flashq::storage::StorageBackend;
-use flashq::storage::file::SyncMode;
 use flashq::{FlashQ, Record};
+use flashq_storage::StorageBackend;
+use flashq_storage::file::SyncMode;
 use test_log::test;
 
 fn make_record(i: usize) -> Record {
