+++ crates/flashq-storage/tests/storage/segment_manager_tests.rs
@@ -1,6 +1,6 @@
 use super::test_utilities::*;
 use flashq::Record;
-use flashq::storage::file::{IndexingConfig, SegmentManager, SyncMode};
+use flashq_storage::file::{IndexingConfig, SegmentManager, SyncMode};
 use test_log::test;
 
 #[test]
