+++ crates/flashq-storage/tests/storage/time_polling_tests.rs
@@ -2,9 +2,9 @@ use super::test_utilities::*;
 use chrono::{DateTime, Duration, FixedOffset};
 use flashq::Record;
 use flashq::RecordWithOffset;
-use flashq::storage::TopicLog;
-use flashq::storage::backend::StorageBackend;
-use flashq::storage::file::{FileTopicLog, SyncMode};
+use flashq_storage::TopicLog;
+use flashq_storage::backend::StorageBackend;
+use flashq_storage::file::{FileTopicLog, SyncMode};
 use std::thread;
 use std::time::Duration as StdDuration;
 use test_log::test;
